backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
aioredis==2.0.1
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.1
httpx==0.26.0
tenacity==8.2.3
qiniu==7.8.0
chardet==5.2.0
python-dotenv==1.0.0
wechatpayv3==1.2.7